+++ lua/terrortown/lang/pt_br.lua
@@ -292,6 +292,7 @@ L.c4_remove_destroy2 = "onfirmar: destruir"
 L.c4_disarm = "Desarmar C4"
 L.c4_disarm_cut = "Clique para cortar o fio {num}"
 
+L.c4_disarm_t     = "Corte um fio para desarmar a bomba. Como você é um Traidor, todos os fios são seguros. Para os Inocentes, porém, não é tão fácil assim!"
 L.c4_disarm_owned = "Corte um fio para desarmar a bomba. É a sua bomba, então todos os fios a desarmarão."
 L.c4_disarm_other = "Corte um fio seguro para desarmar a bomba. Vai explodir se você errar!"
 
@@ -977,8 +978,8 @@ L.target_pickup = "Pressione [{usekey}] para pegar"
 L.target_slot_info = "Slot: {slot}"
 L.target_pickup_weapon = "Pressione [{usekey}] para pegar a arma"
 L.target_switch_weapon = "Pressione [{usekey}] para trocar com a sua arma no momento"
-L.target_pickup_weapon_hidden = ", pressione [{usekey} + {walkkey}] para esconder a arma"
-L.target_switch_weapon_hidden = ", pressione [{usekey} + {walkkey}] para esconder"
+L.target_pickup_weapon_hidden = ", pressione [{walkkey} + {usekey}] para esconder a arma"
+L.target_switch_weapon_hidden = ", pressione [{walkkey} + {usekey}] para esconder"
 L.target_switch_weapon_nospace = "Não há slot de inventário para esta arma"
 L.target_switch_drop_weapon_info = "Dropar {name} do slot {slot}"
 L.target_switch_drop_weapon_info_noslot = "Não há uma arma dropavel no slot {slot}"
@@ -1805,9 +1806,9 @@ L.xfer_team_indicator = "Time"
 --L.label_sherlock_mode = "Enable sherlock mode"
 --L.label_highlight_admins = "Highlight server admins"
 --L.label_highlight_dev = "Highlight TTT2 developer"
---L.label_highlight_vip = "Highlight TTT2 VIP"
+--L.label_highlight_vip = "Highlight TTT2 supporter"
 --L.label_highlight_addondev = "Highlight TTT2 addon developer"
---L.label_highlight_supporter = "Highlight other supporters"
+--L.label_highlight_supporter = "Highlight others"
 --L.label_enable_hud_element = "Enable {elem} HUD element"
 --L.label_spec_prop_control = "Enable prop possession"
 --L.label_spec_prop_base = "Possession base value"
@@ -1866,3 +1867,55 @@ L.xfer_team_indicator = "Time"
 --This value modifies how exponentially fall damage is increased with the speed the player hits the ground at.
 
 --Take care when altering this value. Setting it too high can make even the smallest falls lethal, while setting it too low will allow players to fall from extreme heights and suffer little to no damage.]]
+
+-- 2023-02-08
+--L.testpopup_title = "A Test Popup, now with a multiline title, how NICE!"
+--L.testpopup_subtitle = "Well, hello there! This is a fancy popup with some special information. The text can be also multiline, how fancy! Ugh, I could add so much more text if I'd had any ideas..."
+
+--L.hudeditor_chat_hint1 = "[TTT2][INFO] Hover over an element, press and hold [LMB] and move the mouse to MOVE or RESIZE it."
+--L.hudeditor_chat_hint2 = "[TTT2][INFO] Press and hold the ALT key for symmetric resizing."
+--L.hudeditor_chat_hint3 = "[TTT2][INFO] Press and hold the SHIFT key to move on axis and to keep the aspect ratio."
+--L.hudeditor_chat_hint4 = "[TTT2][INFO] Press [RMB] -> 'Close' to exit the HUD Editor!"
+
+--L.guide_nothing_title = "Nothing here yet!"
+--L.guide_nothing_desc = "This is work in progress, help us by contributing to the project on GitHub."
+
+--L.sb_rank_tooltip_developer = "TTT2 Developer"
+--L.sb_rank_tooltip_vip = "TTT2 Supporter"
+--L.sb_rank_tooltip_addondev = "TTT2 Addon Developer"
+--L.sb_rank_tooltip_admin = "Server Admin"
+--L.sb_rank_tooltip_streamer = "Streamer"
+--L.sb_rank_tooltip_heroes = "TTT2 Heroes"
+--L.sb_rank_tooltip_team = "Team"
+
+--L.tbut_adminarea = "ADMIN AREA:"
+
+-- 2023-08-10
+--L.equipmenteditor_name_damage_scaling = "Damage Scaling"
+
+
+-- 2023-08-11
+--L.equipmenteditor_name_allow_drop = "Allow Drop"
+--L.equipmenteditor_desc_allow_drop = "If enabled, the equipment can be dropped freely by the player."
+
+--L.equipmenteditor_name_drop_on_death_type = "Drop on Death"
+--L.equipmenteditor_desc_drop_on_death_type = "Attempt overriding the action taken for whether the equipment is dropped on player's death."
+
+--L.drop_on_death_type_default = "Default (weapon-defined)"
+--L.drop_on_death_type_force = "Force Drop on Death"
+--L.drop_on_death_type_deny = "Deny Drop on Death"
+
+
+-- 2023-08-26
+--L.equipmenteditor_name_kind = "Equipment Slot"
+--L.equipmenteditor_desc_kind = "The inventory slot the equipment will occupy."
+
+--L.slot_weapon_melee = "Melee Slot"
+--L.slot_weapon_pistol = "Pistol Slot"
+--L.slot_weapon_heavy = "Heavy Slot"
+--L.slot_weapon_nade = "Grenade Slot"
+--L.slot_weapon_carry = "Carry Slot"
+--L.slot_weapon_unarmed = "Unarmed Slot"
+--L.slot_weapon_special = "Special Slot"
+--L.slot_weapon_extra = "Extra Slot"
+--L.slot_weapon_class = "Class Slot"