+++ lua/terrortown/lang/zh_tw.lua
@@ -288,6 +288,7 @@ L.c4_remove_destroy2 = "確認：銷毀"
 L.c4_disarm = "拆除C4"
 L.c4_disarm_cut = "點擊以剪斷 {num} 號引線"
 
+L.c4_disarm_t     = "剪斷引線以拆除C4。您是叛徒，當然每條引線都是安全的，但其他人可就沒那麼容易了！"
 L.c4_disarm_owned = "剪斷引線以拆除C4。你是裝置此C4的人，細節瞭然於胸，任一條引線都可成功拆除。"
 L.c4_disarm_other = "剪斷正確的引線以拆除C4。倘若你犯了錯，後果將不堪設想唷！"
 
@@ -949,8 +950,8 @@ L.target_pickup = "按 [{usekey}] 撿起"
 L.target_slot_info = "槽位：{slot}"
 L.target_pickup_weapon = "按 [{usekey}] 撿起武器"
 L.target_switch_weapon = "按 [{usekey}] 和當前武器交換"
-L.target_pickup_weapon_hidden = "按 [{usekey} + {walkkey}] 隱秘地撿起"
-L.target_switch_weapon_hidden = "按 [{usekey} + {walkkey}] 隱秘地交換"
+L.target_pickup_weapon_hidden = "按 [{walkkey} + {usekey}] 隱秘地撿起"
+L.target_switch_weapon_hidden = "按 [{walkkey} + {usekey}] 隱秘地交換"
 L.target_switch_weapon_nospace = "沒有提供給這個武器的槽位"
 L.target_switch_drop_weapon_info = "丟棄槽位 {slot} 的 {name}"
 L.target_switch_drop_weapon_info_noslot = "槽位 {slot} 沒有可丟棄的武器"
@@ -1540,7 +1541,7 @@ L.kill_score_team = "擊殺隊友："
 -- 2021-10-09
 L.help_models_select = [[
 左鍵點擊模型，將其添加到玩家模型庫中。再次以左鍵刪除它們。右鍵可在所關注的模型的啟用和禁用偵探帽之間進行切換。
-	
+
 左上角的小指示器顯示玩家模型是否有頭部的命中箱，下面的圖標顯示了這個模型是否可佩戴偵探帽。]]
 L.menu_roles_title = "角色設置"
 L.menu_roles_description = "設置生成概率、裝備積分及更多。"
@@ -1779,9 +1780,9 @@ L.label_armor_threshold_for_reinforced = "強化護甲閾值"
 L.label_sherlock_mode = "啟用偵探模式"
 L.label_highlight_admins = "突出服務器管理員"
 L.label_highlight_dev = "突出顯示TTT2開發者"
-L.label_highlight_vip = "高亮顯示 TTT2 VIP"
+L.label_highlight_vip = "突出顯示TTT2支持者"
 L.label_highlight_addondev = "突出顯示TTT2附加組件的開發者"
-L.label_highlight_supporter = "突出顯示其他支持者"
+L.label_highlight_supporter = "突出顯示其他人"
 L.label_enable_hud_element = "啟用{elem}HUD元素"
 L.label_spec_prop_control = "啟用Prop附體"
 L.label_spec_prop_base = "附體時的基礎值"
@@ -1843,3 +1844,53 @@ L.help_falldmg_exponent = [[
 該值修改了隨著玩家撞擊地面的速度而以指數方式增加的摔落傷害。
 
 更改此值時請小心。設置得太高，即使是二階階梯的高度也會致命，而設置得太低，玩家從五樓跳下來時也安然無恙。]]
+
+-- 2023-02-08
+L.testpopup_title = "一個測試彈出窗口，現在有一個多行標題，多好啊！"
+L.testpopup_subtitle = "好吧，你好！這是一個帶有一些特殊信息的花式彈出窗口。文字也可以是多行的，多好啊！呃，如果我有什麼想法的話，我可以添加這麼多的文字..."
+
+L.hudeditor_chat_hint1 = "[TTT2][INFO] 將鼠標懸停在一個元素上，按住[LMB]並移動鼠標來移動或調整其大小。"
+L.hudeditor_chat_hint2 = "[TTT2][INFO] 按住ALT鍵進行對稱調整大小。"
+L.hudeditor_chat_hint3 = "[TTT2][INFO] 按住SHIFT鍵，在軸上移動並保持長寬比。"
+L.hudeditor_chat_hint4 = "[TTT2][INFO] 按[RMB] -> 'Close' 來退出HUD編輯器！"
+
+L.guide_nothing_title = "這裡暫時什麼都沒有！"
+L.guide_nothing_desc = "這是一項正在進行中的工作，通過在GitHub上為項目做貢獻來幫助我們。"
+
+L.sb_rank_tooltip_developer = "TTT2開發者"
+L.sb_rank_tooltip_vip = "TTT2支持者"
+L.sb_rank_tooltip_addondev = "TTT2附加組件開發者"
+L.sb_rank_tooltip_admin = "服務器管理員"
+L.sb_rank_tooltip_streamer = "主播"
+L.sb_rank_tooltip_heroes = "TTT2 英雄"
+L.sb_rank_tooltip_team = "陣營"
+
+L.tbut_adminarea = "管理區:"
+
+-- 2023-08-10
+L.equipmenteditor_name_damage_scaling = "傷害縮放"
+
+-- 2023-08-11
+L.equipmenteditor_name_allow_drop = "允許丟棄"
+L.equipmenteditor_desc_allow_drop = "如果啟用，玩家可以自由地丟棄裝備。"
+
+L.equipmenteditor_name_drop_on_death_type = "死亡時丟棄"
+L.equipmenteditor_desc_drop_on_death_type = "嘗試覆蓋玩家死亡時裝備是否被丟棄的操作。"
+
+L.drop_on_death_type_default = "默認（由武器定義）"
+L.drop_on_death_type_force = "強制死亡時丟棄"
+L.drop_on_death_type_deny = "拒絕死亡時丟棄"
+
+-- 2023-08-26
+L.equipmenteditor_name_kind = "裝備槽"
+L.equipmenteditor_desc_kind = "裝備將佔用的庫存槽。"
+
+L.slot_weapon_melee = "近戰槽"
+L.slot_weapon_pistol = "手槍槽"
+L.slot_weapon_heavy = "重型槽"
+L.slot_weapon_nade = "手雷槽"
+L.slot_weapon_carry = "攜帶槽"
+L.slot_weapon_unarmed = "空手槽"
+L.slot_weapon_special = "特殊槽"
+L.slot_weapon_extra = "額外槽"
+L.slot_weapon_class = "職業槽"